+++ platformio/fs.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import glob
 import hashlib
 import io
 import json
@@ -24,7 +25,7 @@ import sys
 import click
 
 from platformio import exception
-from platformio.compat import WINDOWS, glob_escape, glob_recursive
+from platformio.compat import WINDOWS
 
 
 class cd(object):
@@ -158,7 +159,9 @@ def match_src_files(src_dir, src_filter=None, src_exts=None, followlinks=True):
     src_filter = src_filter.replace("/", os.sep).replace("\\", os.sep)
     for (action, pattern) in re.findall(r"(\+|\-)<([^>]+)>", src_filter):
         items = set()
-        for item in glob_recursive(os.path.join(glob_escape(src_dir), pattern)):
+        for item in glob.glob(
+            os.path.join(glob.escape(src_dir), pattern), recursive=True
+        ):
             if os.path.isdir(item):
                 for root, _, files in os.walk(item, followlinks=followlinks):
                     for f in files: